--- a/lib/optimize.js
+++ b/lib/optimize.js
@@ -27,8 +27,8 @@ function optimizeJsonConcat (code) {
 
   for (let i = 0; i < code.length; i++) {
     if (i > 0 && jsonConcatRE.test(code[i]) && jsonConcatRE.test(code[i - 1])) {
-      const mergedEntry = code[i - 1] + ' +' + code[i].substring(code[i].indexOf('json +=') + 7)
-      optimizedJsonConcat.pop() // Remove the previous entry
+      const lastEntry = optimizedJsonConcat.pop()
+      const mergedEntry = lastEntry + ' +' + code[i].substring(code[i].indexOf('json +=') + 7)
       optimizedJsonConcat.push(mergedEntry)
     } else {
       optimizedJsonConcat.push(code[i])