--- a/lib/optimize.js
+++ b/lib/optimize.js
@@ -38,7 +38,7 @@ function optimizeJsonConcat (code) {
   return optimizedJsonConcat
 }
 
-function optimizeLetJson(code) {
+function optimizeLetJson (code) {
   const optimizedLetJsonCode = []
   for (let i = 0; i < code.length; i++) {
     if (i > 0 && jsonConcatRE.test(code[i]) && letJsonRE.test(code[i - 1])) {
@@ -52,7 +52,7 @@ function optimizeLetJson(code) {
   return optimizedLetJsonCode
 }
 
-function optimizeDirectReturn(code) {
+function optimizeDirectReturn (code) {
   const optimizedDirectReturnCode = []
   for (let i = 0; i < code.length; i++) {
     if (i > 0 && returnJsonRE.test(code[i]) && letJsonRE.test(code[i - 1])) {
@@ -66,7 +66,7 @@ function optimizeDirectReturn(code) {
   return optimizedDirectReturnCode
 }
 
-function optimizeReturnEmptyString(code) {
+function optimizeReturnEmptyString (code) {
   for (let i = 0; i < code.length; i++) {
     if (returnEmptyStringRE.test(code[i])) {
       code[i] = code[i].replace('return \'\' +', 'return')