--- a/lib/serializer.js
+++ b/lib/serializer.js
@@ -3,51 +3,9 @@
 // eslint-disable-next-line
 const STR_ESCAPE = /[\u0000-\u001f\u0022\u005c\ud800-\udfff]|[\ud800-\udbff](?![\udc00-\udfff])|(?:[^\ud800-\udbff]|^)[\udc00-\udfff]/
 
-module.exports = class Serializer {
+class Serializer {
   constructor (options) {
-    switch (options && options.rounding) {
-      case 'floor':
-        this.parseInteger = Math.floor
-        break
-      case 'ceil':
-        this.parseInteger = Math.ceil
-        break
-      case 'round':
-        this.parseInteger = Math.round
-        break
-      case 'trunc':
-      default:
-        this.parseInteger = Math.trunc
-        break
-    }
     this._options = options
-  }
-
-  asInteger (i) {
-    if (typeof i === 'number') {
-      if (i === Infinity || i === -Infinity) {
-        throw new Error(`The value "${i}" cannot be converted to an integer.`)
-      }
-      if (Number.isInteger(i)) {
-        return '' + i
-      }
-      if (Number.isNaN(i)) {
-        throw new Error(`The value "${i}" cannot be converted to an integer.`)
-      }
-      return this.parseInteger(i)
-    } else if (i === null) {
-      return '0'
-    } else if (typeof i === 'bigint') {
-      return i.toString()
-    } else {
-      /* eslint no-undef: "off" */
-      const integer = this.parseInteger(i)
-      if (Number.isFinite(integer)) {
-        return '' + integer
-      } else {
-        throw new Error(`The value "${i}" cannot be converted to an integer.`)
-      }
-    }
   }
 
   asNumber (i) {
@@ -113,11 +71,12 @@ module.exports = class Serializer {
       }
     }
 
+    // Fast escape chars check
     if (str.length < 42) {
-      return this.asStringSmall(str)
+      return Serializer.asStringSmall(str)
     } else if (STR_ESCAPE.test(str) === false) {
       return '"' + str + '"'
-    } else {
+    } else  {
       return JSON.stringify(str)
     }
   }
@@ -128,7 +87,7 @@ module.exports = class Serializer {
   // every string that contain surrogate needs JSON.stringify()
   // 34 and 92 happens all the time, so we
   // have a fast case for them
-  asStringSmall (str) {
+  static asStringSmall (str) {
     const len = str.length
     let result = ''
     let last = -1
@@ -161,7 +120,140 @@ module.exports = class Serializer {
     return this._options
   }
 
-  static restoreFromState (state) {
-    return new Serializer(state)
+}
+
+class SerializerFloor extends Serializer {
+  asInteger (i) {
+    if (typeof i === 'number') {
+      if (i === Infinity || i === -Infinity) {
+        throw new Error(`The value "${i}" cannot be converted to an integer.`)
+      }
+      if (Number.isInteger(i)) {
+        return '' + i
+      }
+      if (Number.isNaN(i)) {
+        throw new Error(`The value "${i}" cannot be converted to an integer.`)
+      }
+      return '' + Math.floor(i)
+    } else if (i === null) {
+      return '0'
+    } else if (typeof i === 'bigint') {
+      return i.toString()
+    } else {
+      /* eslint no-undef: "off" */
+      const integer = Math.floor(i)
+      if (Number.isFinite(integer)) {
+        return '' + integer
+      } else {
+        throw new Error(`The value "${i}" cannot be converted to an integer.`)
+      }
+    }
   }
+}
+
+class SerializerCeil extends Serializer {
+  asInteger (i) {
+    if (typeof i === 'number') {
+      if (i === Infinity || i === -Infinity) {
+        throw new Error(`The value "${i}" cannot be converted to an integer.`)
+      }
+      if (Number.isInteger(i)) {
+        return '' + i
+      }
+      if (Number.isNaN(i)) {
+        throw new Error(`The value "${i}" cannot be converted to an integer.`)
+      }
+      return '' + Math.ceil(i)
+    } else if (i === null) {
+      return '0'
+    } else if (typeof i === 'bigint') {
+      return i.toString()
+    } else {
+      /* eslint no-undef: "off" */
+      const integer = Math.ceil(i)
+      if (Number.isFinite(integer)) {
+        return '' + integer
+      } else {
+        throw new Error(`The value "${i}" cannot be converted to an integer.`)
+      }
+    }
+  }
+}
+
+class SerializerTrunc extends Serializer {
+  asInteger (i) {
+    if (typeof i === 'number') {
+      if (i === Infinity || i === -Infinity) {
+        throw new Error(`The value "${i}" cannot be converted to an integer.`)
+      }
+      if (Number.isInteger(i)) {
+        return '' + i
+      }
+      if (Number.isNaN(i)) {
+        throw new Error(`The value "${i}" cannot be converted to an integer.`)
+      }
+      return '' + Math.trunc(i)
+    } else if (i === null) {
+      return '0'
+    } else if (typeof i === 'bigint') {
+      return i.toString()
+    } else {
+      /* eslint no-undef: "off" */
+      const integer = Math.trunc(i)
+      if (Number.isFinite(integer)) {
+        return '' + integer
+      } else {
+        throw new Error(`The value "${i}" cannot be converted to an integer.`)
+      }
+    }
+  }
+}
+
+class SerializerRound extends Serializer {
+  asInteger (i) {
+    if (typeof i === 'number') {
+      if (i === Infinity || i === -Infinity) {
+        throw new Error(`The value "${i}" cannot be converted to an integer.`)
+      }
+      if (Number.isInteger(i)) {
+        return '' + i
+      }
+      if (Number.isNaN(i)) {
+        throw new Error(`The value "${i}" cannot be converted to an integer.`)
+      }
+      return '' + Math.round(i)
+    } else if (i === null) {
+      return '0'
+    } else if (typeof i === 'bigint') {
+      return i.toString()
+    } else {
+      /* eslint no-undef: "off" */
+      const integer = Math.round(i)
+      if (Number.isFinite(integer)) {
+        return '' + integer
+      } else {
+        throw new Error(`The value "${i}" cannot be converted to an integer.`)
+      }
+    }
+  }
+}
+
+function SerializerFactory (options) {
+  switch (options && options.rounding) {
+    case 'floor':
+      return new SerializerFloor(options)
+    case 'ceil':
+      return new SerializerCeil(options)
+    case 'round':
+      return new SerializerRound(options)
+    case 'trunc':
+    default:
+      return new SerializerTrunc(options)
+  }
+}
+
+module.exports = SerializerFactory
+module.exports.Serializer = Serializer
+module.exports.restoreFromState = function restoreFromState (state) {
+  return SerializerFactory(state)
 }