--- a/lib/serializer.js
+++ b/lib/serializer.js
@@ -76,7 +76,7 @@ class Serializer {
       return Serializer.asStringSmall(str)
     } else if (STR_ESCAPE.test(str) === false) {
       return '"' + str + '"'
-    } else  {
+    } else {
       return JSON.stringify(str)
     }
   }
@@ -119,7 +119,6 @@ class Serializer {
   getState () {
     return this._options
   }
-
 }
 
 class SerializerFloor extends Serializer {