--- a/test/debug-mode.test.js
+++ b/test/debug-mode.test.js
@@ -7,7 +7,7 @@ const Ajv = require('ajv').default
 const Validator = require('../lib/validator')
 const { Serializer } = require('../lib/serializer')
 
-function build(opts) {
+function build (opts) {
   return fjs({
     title: 'default string',
     type: 'object',