@mcollina 2023-07-26
how much perf would this get us?