@Uzlopak 2023-07-27
Odd:

before:
```
short string............................................. x 43,029,750 ops/sec ±0.82% (191 runs sampled)
short string with double quote........................... x 13,328,954 ops/sec ±0.33% (189 runs sampled)
long string without double quotes........................ x 26,316 ops/sec ±0.27% (192 runs sampled)
long string.............................................. x 11,696 ops/sec ±0.17% (192 runs sampled)
number................................................... x 1,354,808,191 ops/sec ±0.11% (190 runs sampled)
integer.................................................. x 222,011,571 ops/sec ±0.08% (193 runs sampled)
formatted date-time...................................... x 2,838,326 ops/sec ±0.22% (192 runs sampled)
formatted date........................................... x 1,595,951 ops/sec ±0.23% (191 runs sampled)
formatted time........................................... x 1,603,617 ops/sec ±0.22% (190 runs sampled)
short array of numbers................................... x 75,758 ops/sec ±0.36% (189 runs sampled)
short array of integers.................................. x 68,997 ops/sec ±0.46% (190 runs sampled)
short array of short strings............................. x 22,570 ops/sec ±0.28% (189 runs sampled)
short array of long strings.............................. x 22,504 ops/sec ±0.25% (188 runs sampled)
short array of objects with properties of different types x 11,132 ops/sec ±0.33% (189 runs sampled)
object with number property.............................. x 1,341,001,463 ops/sec ±0.13% (188 runs sampled)
object with integer property............................. x 221,718,067 ops/sec ±0.11% (193 runs sampled)
object with short string property........................ x 41,396,300 ops/sec ±0.69% (189 runs sampled)
object with long string property......................... x 11,630 ops/sec ±0.19% (192 runs sampled)
object with properties of different types................ x 2,129,851 ops/sec ±0.55% (187 runs sampled)
simple object............................................ x 9,928,667 ops/sec ±0.44% (190 runs sampled)
object with const string property........................ x 1,354,504,033 ops/sec ±0.06% (194 runs sampled)
object with const number property........................ x 1,356,175,724 ops/sec ±0.07% (194 runs sampled)
object with const bool property.......................... x 1,349,457,006 ops/sec ±0.13% (191 runs sampled)
object with const object property........................ x 1,352,233,428 ops/sec ±0.14% (193 runs sampled)
object with const null property.......................... x 1,356,402,514 ops/sec ±0.08% (193 runs sampled)

```
after
```
short string............................................. x 42,957,374 ops/sec ±0.53% (191 runs sampled)
short string with double quote........................... x 13,084,788 ops/sec ±0.74% (186 runs sampled)
long string without double quotes........................ x 26,446 ops/sec ±0.32% (191 runs sampled)
long string.............................................. x 11,737 ops/sec ±0.17% (192 runs sampled)
number................................................... x 1,361,444,317 ops/sec ±0.07% (192 runs sampled)
integer.................................................. x 222,455,101 ops/sec ±0.08% (194 runs sampled)
formatted date-time...................................... x 2,961,085 ops/sec ±0.18% (189 runs sampled)
formatted date........................................... x 1,642,124 ops/sec ±0.47% (191 runs sampled)
formatted time........................................... x 1,637,101 ops/sec ±0.20% (190 runs sampled)
short array of numbers................................... x 75,280 ops/sec ±0.35% (187 runs sampled)
short array of integers.................................. x 69,451 ops/sec ±0.43% (190 runs sampled)
short array of short strings............................. x 22,055 ops/sec ±0.30% (190 runs sampled)
short array of long strings.............................. x 23,073 ops/sec ±0.43% (191 runs sampled)
short array of objects with properties of different types x 10,003 ops/sec ±0.28% (190 runs sampled)
object with number property.............................. x 1,345,255,460 ops/sec ±0.13% (187 runs sampled)
object with integer property............................. x 224,592,369 ops/sec ±0.06% (193 runs sampled)
object with short string property........................ x 42,360,931 ops/sec ±0.23% (191 runs sampled)
object with long string property......................... x 11,612 ops/sec ±0.19% (192 runs sampled)
object with properties of different types................ x 1,768,995 ops/sec ±0.47% (188 runs sampled)
simple object............................................ x 12,824,498 ops/sec ±0.52% (190 runs sampled)
object with const string property........................ x 1,361,069,070 ops/sec ±0.08% (194 runs sampled)
object with const number property........................ x 1,360,204,001 ops/sec ±0.11% (189 runs sampled)
object with const bool property.......................... x 1,354,428,630 ops/sec ±0.06% (194 runs sampled)
```

simple object is faster and I kind of can break the perf of typia. but "object with properties of different types" is slower.