@cesco69 2024-03-01
@Uzlopak this PR looks very interesting, why not keep working on it?